Move setupM into internal; fix cycleTx
