+++ cardano-server.cabal
@@ -38,7 +38,6 @@ library
       Server.Endpoints.Ping
       Server.Internal
       Server.Main
-      Server.Setup
       Server.Tx
       TestingServer.Main
       TestingServer.OnChain
@@ -104,7 +103,6 @@ test-suite cardano-server-test
     plutus-ledger                          -any,
     text                                   -any,
     time                                   -any
-
 
 executable testingServer
   default-language: Haskell2010
